microservices/notification_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
redis==5.0.1
django-redis==5.4.0
requests==2.31.0
pydantic==2.5.0
django-environ==0.11.2
gunicorn==21.2.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
django-stubs==4.2.7
pytest-cov==4.1.0
django-health-check==3.17.0
prometheus-client==0.19.0
structlog==23.2.0
django-structlog==6.0.0
cryptography==41.0.7
PyJWT==2.8.0
# Email and SMS specific dependencies
django-anymail==10.2
twilio==8.10.0
sendgrid==6.10.0
jinja2==3.1.2
premailer==3.10.0
beautifulsoup4==4.12.2
lxml==4.9.3